+++ app/src/main/java/com/techcourse/AppWebApplicationInitializer.java
@@ -3,7 +3,7 @@ package com.techcourse;
 import jakarta.servlet.ServletContext;
 import nextstep.mvc.DispatcherServlet;
 import nextstep.mvc.controller.tobe.handleradapter.AnnotationHandlerAdapter;
-import nextstep.mvc.controller.tobe.handleradapter.ManualHandlerAdapter;
+import nextstep.mvc.controller.asis.ManualHandlerAdapter;
 import nextstep.mvc.controller.tobe.handlermapping.AnnotationHandlerMapping;
 import nextstep.web.WebApplicationInitializer;
 import org.slf4j.Logger;
